cctv-backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
aiofiles
numpy
torch
torchvision
torchaudio
librosa
imageio-ffmpeg
opencv-python
Pillow
mediapipe
